# Avoid object allocation when creating strings from the underlying buffer