For me this makes a significant difference WRT decoding performance, since fewer short lived copies of ~strings~buffers are being created.

See https://gist.github.com/rklaehn/ba52f94896788a67107e3d38fc37a498