@rklaehn 2018-10-08
Here are the builtin benchmarks:

master:
```
PID: 677
encode - node-cbor - 76 x 622 ops/sec ±0.83% (74 runs sampled)
encode - borc - 76 x 15,305 ops/sec ±3.40% (82 runs sampled)
encode - stream - borc - 76 x 5,821 ops/sec ±7.12% (74 runs sampled)
encode - JSON.stringify - 76 x 27,993 ops/sec ±7.05% (74 runs sampled)
decode - node-cbor - 47 x 1,453 ops/sec ±1.27% (82 runs sampled)
decode - borc - 47 x 14,708 ops/sec ±2.31% (79 runs sampled)
decode - JSON.parse - 47 x 49,168 ops/sec ±6.08% (65 runs sampled)
```

this branch:
```
PID: 322
encode - node-cbor - 76 x 581 ops/sec ±1.28% (76 runs sampled)
encode - borc - 76 x 15,611 ops/sec ±3.07% (84 runs sampled)
encode - stream - borc - 76 x 6,051 ops/sec ±3.80% (79 runs sampled)
encode - JSON.stringify - 76 x 27,777 ops/sec ±7.35% (73 runs sampled)
decode - node-cbor - 47 x 1,509 ops/sec ±1.40% (82 runs sampled)
decode - borc - 47 x 31,221 ops/sec ±3.93% (82 runs sampled)
decode - JSON.parse - 47 x 50,254 ops/sec ±6.10% (65 runs sampled)
```