@rklaehn 2018-10-18
No problem. Thanks for merging this. I want to move from JSON to CBOR for some use case, and CBOR not being much slower to parse will definitely help with that tradeoff...