@dignifiedquire 2018-10-18
Published as 2.04

Yeah bring more cbor to the world!
On 18. Oct 2018, 15:13 +0200, Rüdiger Klaehn <notifications@github.com>, wrote:
> No problem. Thanks for merging this. I want to move from JSON to CBOR for some use case, and CBOR not being much slower to parse will definitely help with that tradeoff...
> —
> You are receiving this because you modified the open/close state.
> Reply to this email directly, view it on GitHub, or mute the thread.